+++ app/services/bot_context_v2/commands/check.rb
@@ -18,6 +18,7 @@ module BotContextV2
         when 'Fate::Character' then BotContextV2::Commands::Checks::Fate.new
         when 'Fallout::Character' then BotContextV2::Commands::Checks::Fallout.new
         when 'Cthulhu7::Character' then BotContextV2::Commands::Checks::Cthulhu7.new
+        when 'Nimble::Character' then BotContextV2::Commands::Checks::Nimble.new
         end
       end
     end