+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - attacks for nimble characters
 - feats for nimble characters
 - leveling for nimble characters
+- rolls for nimble characters
 
 ## [0.5.2] - 2026-07-24
 ### Added